requirements.txt
openai>=1.0.0
pyautogui>=0.9.54
mss>=9.0.0
pillow>=10.0.0
